service/socialnetwork/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-allauth-ui==0.2.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-taggit==5.0.1
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
